requirements.txt
streamlit
pypdf2
langchain
langchain-community
faiss-cpu
huggingface_hub
python-dotenv
InstructorEmbedding
sentence_transformers